Make public/bundles/nelmioapidoc a real target, during build
This makes sure it is copied (locally, in the source tree)
as the current user, preventing permission issues cleaning it
up again. During install it is copied with everything else to
the target tree.

By making it a real, non-PHONY target, it does not get recreated
again during `make install-docs`, which could then also likely
be with root ownership.